tool_poc/requirements.txt
python-dotenv>=1.0.0
click>=8.1.7
openai>=1.10.0
pinecone>=3.0.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1 